replaced parseInt with parseFloat

+++ js/javascript.js
@@ -1,10 +1,10 @@
 function triangleChecker(){
   var height = document.getElementById("length1").value;
-  var newHeight = parseInt(height);
+  var newHeight = parseFloat(height);
   var length = document.getElementById("length2").value;
-  var newLength = parseInt(length);
+  var newLength = parseFloat(length);
   var width = document.getElementById("length3").value;
-  var newWidth = parseInt(width);
+  var newWidth = parseFloat(width);
   var triangle = [];
 
   triangle = [newHeight,newLength,newWidth];
